General Layout
I put together a layout that looks similar to Etherscan.  I tried to get the general look of the website replicated. This seems to look alright. It has a navbar at the top with categories, general info about ETH, and the latest blocks/transactions. I may add search functionality to search blocks and transactions.